Merged PR 78322: VSTS #297411 - Check for remapped level ids on an update.
EWR updated an embedded reference file's level map causing the level ids of the levels to be changed.  The converter previously was only looking for an existing category by level id.  Now it checks to make sure the name of the level is consistent.  If not, it will look for an XSA with that name and then update the mapping.  If level ids match, but the name was previously unused, the converter will create a new category.

Related work items: #297411