Merged PR 78622: #297134 Improve performance snapping to large parts.
#297134 Improve performance snapping to large parts. Fix snap returning obscured edge for multi-primitive geometry streams.

In addition to adding the sub-range cache for multi-entry parts, this fixes a significant issue with snapping to geometry streams containing multiple surface primitives. Previously the snap path returned would be the first curve found within the snap aperture without considering the z distance from the test point. This is only correct for a single surface primitive as we never test it's obscured edges, but in the multi-surface case, the edge can be obscured by a different surface.

Related work items: #297134