Merged PR 79143: update must also use ref naming rule
I had the new rule for naming a referenced model only in the logic that creates a model for the first time. The rule must also be applied when its reference attachment name changes. (In both cases, the name of a referenced model is the name of the attachment, which is based partly on the name of the referenced file. So the name of a referenced model changes when the name of the referenced file changes. And, of course, it changes when the reference moves from one parent to another.)
I also added some addition trace-level logging to iModelBridgeFwk to help us debug why a customer sees bridges freeze in iTwin Synchronizer.

Related work items: #301690