Merged PR 80422: If the holder doesn't exist, then don't do anything on a reset.
Michel D. reported that ConversionCustomAttributeHelper::Reset was causing an exception to be thrown.  It seems like the Reset() was causing a new Holder to be created (which is dumb) and that was failing, so the call to Reset() was on a null pointer.  So now just check to see if we have a holder and if not, don't do anything.