Merged PR 83702: Must upgrade registry to allow OBD bridge to add assignments
The registry schema has changes. BAS is creating the db with the old schema, and OBD bridge is trying to write with the new. So, we must upgrade the db when we open it. This will be harmless to BAS and any old bridges that may  be running, since we only add columns and tables.