Merged PR 85670: Enforce the rule that a bridge must NOT call DgnDb::SaveChanges directly
If a bridge calls DgnDb::SaveChanges directly, that terminates and does not restart the bulk operation. That in turn leads to hard-to-debug problems later on (for example, a primitive insert fails because locks and codes are not held). The bridge must instead call iModelBridge::SaveChanges[ToConserveMemory]. This change enforces this rule.

I also found that iModelBridgeFwk was returning an uninitialized value as its exit code. That is because iModelBridgeError was not initializing its m_id member, and lots of code fails to set it. We need an extensive sweep through fwk to clean this up.