Merged PR 85184: handle failure to capture changeset
commit must fail if changeset cannot be created
bridge must fail if commit fails
SQLite imposes a hard limit of 2GB to the size of a changeset (session data).
The underlying problem is that that bridge should save and commit more frequently.
In the test case, there were many drawing models in a single file, and the bridge did not commit/push in the course of converting them but only at the end.