Merged PR 87308: Tweaking the logic calculating affinity, giving priority to the dgn's authoring prod-name.
Tweaking the logic calculating affinity, giving priority to the dgn's authoring prod-name. Disabled DiscloseAndAffinities test since it can't run in the same process as the new GetAffinityTest (iModelBridgeFwk tries to initialize the DgnV8 Lib in both).

In summary, the new affinity logic is:

if (authoring_prod_name is one of the OpenXXX products)
    return ExactMatch

if (file is Civilized)
    {
    if (file has no authoring_prod_name)
        return ExactMatch
    else
        return HighAffinity
    }

return LowAffinity