Merged PR 89668: Report level diffs sooner and more completely
The MicroStation bridge should report inconsistent level definitions when they are first encountered and should not wait until a given file is processed the second time.
It should also report all of the properties of a level's definition that differ from the standard one and not just the first difference.
The user (Bentley consultant? Bentley cert?) thinks this will be more helpful.

Related work items: #354139