Merged PR 88774: setup dll path for OBD bridge (affinity)
The call to discloseFilesAndAffinities creates an instance of the bridge and initializes it. Initialization needs to load V8 DLLs, and so the path to those DLLs must be set up. OBD bridge has some of the V8 DLLs in the main directory and some in subdirectories. The only one that seems to be needed by the bridge is MdlSys/AsNeeded (for TerrainModelElementHandler.dll).
Only a PP-based bridge like OBD can have this non-standard setup. So, this hack should be sufficient.