Merged PR 89122: add classes to access BlobDaemon
This adds support for starting and accessing the BlobDaemon from JavaScript. There are also changes to the package.json for the imodeljs-native package, so compensating changes are required in imodeljs-backend in [this imodeljs PR](https://bentleycs.visualstudio.com/iModelTechnologies/_git/imodeljs/pullrequest/89809)

The BlobDaemon.exe is delivered next to the imodeljs.node file so it can be located and started from JavaScript.

> Note that this PR also adds new dependencies in the BeSQLite part on Curl and OpenSSL. Hopefully that won't negatively impact other consumers of that part.