Merged PR 90164: GeometricModel can override ViewFlags
GeoBridge users produce models with lots of patterned areas that are intended to be viewed in wireframe - they are unintelligible in smooth shaded mode. The symbology is mandated by strict standards.
They also want to produce views that include not only these types of models, but others (architectural etc), and they want those models to be displayed in smooth shaded mode.
Added a `viewFlagOverrides` JSON property to GeometricModel. At display time, we will apply those overrides to the view's ViewFlags.

Updated the native ViewFlags and ViewFlagOverrides to match the current TypeScript versions.
Added ViewFlagsOverrides::To/FromJson().

Corresponding TypeScript PR forthcoming...