Merged PR 91022: Support update of SectionLocation element
The SectionLocation conversion logic can be handled like a normal element conversion. It is a conversion from a V8 element (NamedView complex element) to a BIM element. So, we can use the ChangeDetector and other standard methods to check for changes and maintain an ExternalSourceAspect. The only trick was to make it possible to map the V8 dictionary model to the job's definition model, since that is the source of the V8 NamedView element.

Also fixed bug in ConvertModelACSTraverser where it was using the wrong scope element when checking for existing ACS element in iModel.