Fix JS logging leading to a crash on mobile (#568)
* Fix JS logging leading to a crash on mobile
On mobile devices where backend logging is configured to forward to the frontend, doing a JS log too early in the startup process results in an unhandled exception due to the fact that the frontend and backend are not yet connected.
Catch that exception when it happens and store the failed log message in the list of deferred log messages.
Also process the list of deferred log messages any time IsSeverityEnabled is called so that you don't have to wait for an enabled severity message to be logged before the deferred ones get processed.

* Improve comments explaining exception handlers

* Revert accidental commit.

* Improve comments explaining exception handlers

* Fix reversion

* Remove deferred processing in IsSeverityEnabled
Required by Keith Bentley

(cherry picked from commit 5b96ccda41af1b8caa12745ab8947912a9578d7a)