Make SchemaMerger more robust (#644)
Solves the issue of OB Connector crashing when a new merger is used, as the schemas had wrong versions. This PR allows to bypass the unwanted behavior by setting a DoNotMergeReferences flag. The flag enables not inserting missing schema references and copying source references instead of locating it. Additionally, schema lists validation and null checks were added for the merge operation to make it more robust and more descriptive in case of an error.

Co-authored-by: Robert Schili <rschili@users.noreply.github.com>