Fix Xcode 15.4 errors (#837)
* Fix Xcode 15.4 errors
Note: the original errors had to do with comparing a RefCountedPtr to nullptr. The only fix I could come up with for that caused comparisons to NULL to then fail, so I updated those places to instead use nullptr. There are other changes that I had to make that shouldn't have been Mac-specific. Specifically, the changes in blockcachevfsd.c and t_normals.cpp.

* More NULL -> nullptr updates

* fix pre-existing logic error in test

* Update iModelCore/BeSQLite/SQLite/blockcachevfsd.c

---------

Co-authored-by: dassaf4 <68340676+dassaf4@users.noreply.github.com>
Co-authored-by: Nick Tessier <22119573+nick4598@users.noreply.github.com>